+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [0.2.2] - 2025-07-24
+- Added items tab
+- Added select and unselect all buttons in the add recipe modal in the recipes tab
+- Update LP-Optimize to only optimize outputs with targets to prevent computing for infinity
+
 ## [0.1.1] - 2025-07-22
 - Added satisfactory data set
 - Added numerous changes to unify factorio and satisfactory data handling